requirements.txt
plyer
pyserial
dbus-python
argparse
pygame
